Finished basic routing
All basic routing done.
2 main types root and auth
We will render navs and footers in roout, whereas no need in auth

we have created routes for applying functionalities.
user can make use of the functionalities iff he has enough of the credits, so credits page has been created.
for the transformations part, this route is only accesssible if contined with id, then it can be updated,
type can be added to transformations